--- lib/datadog/core/transport/http/api/instance.rb
@@ -1,37 +1,0 @@
-# frozen_string_literal: true
-
-module Datadog
-  module Core
-    module Transport
-      module HTTP
-        module API
-          # An API configured with adapter and routes
-          class Instance
-            attr_reader \
-              :adapter,
-              :headers,
-              :spec
-
-            def initialize(spec, adapter, options = {})
-              @spec = spec
-              @adapter = adapter
-              @headers = options.fetch(:headers, {})
-            end
-
-            def encoder
-              spec.encoder
-            end
-
-            def call(env)
-              # Add headers to request env, unless empty.
-              env.headers.merge!(headers) unless headers.empty?
-
-              # Send request env to the adapter.
-              adapter.call(env)
-            end
-          end
-        end
-      end
-    end
-  end
-end
